Module aws_sdk_workdocs::error

Modules

concurrent_modification_exception

See ConcurrentModificationException

conflicting_operation_exception

See ConflictingOperationException

custom_metadata_limit_exceeded_exception

See CustomMetadataLimitExceededException

deactivating_last_system_user_exception

See DeactivatingLastSystemUserException

document_locked_for_comments_exception

See DocumentLockedForCommentsException

draft_upload_out_of_sync_exception

See DraftUploadOutOfSyncException

entity_already_exists_exception

See EntityAlreadyExistsException

entity_not_exists_exception

See EntityNotExistsException

failed_dependency_exception

See FailedDependencyException

illegal_user_state_exception

See IllegalUserStateException

invalid_argument_exception

See InvalidArgumentException

invalid_comment_operation_exception

See InvalidCommentOperationException

invalid_operation_exception

See InvalidOperationException

invalid_password_exception

See InvalidPasswordException

limit_exceeded_exception

See LimitExceededException

prohibited_state_exception

See ProhibitedStateException

requested_entity_too_large_exception

See RequestedEntityTooLargeException

resource_already_checked_out_exception

See ResourceAlreadyCheckedOutException

service_unavailable_exception

See ServiceUnavailableException

storage_limit_exceeded_exception

See StorageLimitExceededException

storage_limit_will_exceed_exception

See StorageLimitWillExceedException

too_many_labels_exception

See TooManyLabelsException

too_many_subscriptions_exception

See TooManySubscriptionsException

unauthorized_operation_exception

See UnauthorizedOperationException

unauthorized_resource_access_exception

See UnauthorizedResourceAccessException

Structs

AbortDocumentVersionUploadError

ActivateUserError

AddResourcePermissionsError

ConcurrentModificationException

The resource hierarchy is changing.

ConflictingOperationException

Another operation is in progress on the resource that conflicts with the current operation.

CreateCommentError

CreateCustomMetadataError

CreateFolderError

CreateLabelsError

CreateNotificationSubscriptionError

CreateUserError

CustomMetadataLimitExceededException

The limit has been reached on the number of custom properties for the specified

DeactivateUserError

DeactivatingLastSystemUserException

The last user in the organization is being deactivated.

DeleteCommentError

DeleteCustomMetadataError

DeleteDocumentError

DeleteFolderContentsError

DeleteFolderError

DeleteLabelsError

DeleteNotificationSubscriptionError

DeleteUserError

DescribeActivitiesError

DescribeCommentsError

DescribeDocumentVersionsError

DescribeFolderContentsError

DescribeGroupsError

DescribeNotificationSubscriptionsError

DescribeResourcePermissionsError

DescribeRootFoldersError

DescribeUsersError

DocumentLockedForCommentsException

This exception is thrown when the document is locked for comments and user tries to

DraftUploadOutOfSyncException

This exception is thrown when a valid checkout ID is not presented on document

EntityAlreadyExistsException

The resource already exists.

EntityNotExistsException

The resource does not exist.

FailedDependencyException

The AWS Directory Service cannot reach an on-premises instance. Or a dependency

GetCurrentUserError

GetDocumentError

GetDocumentPathError

GetDocumentVersionError

GetFolderError

GetFolderPathError

GetResourcesError

IllegalUserStateException

The user is undergoing transfer of ownership.

InitiateDocumentVersionUploadError

InvalidArgumentException

The pagination marker or limit fields are not valid.

InvalidCommentOperationException

The requested operation is not allowed on the specified comment object.

InvalidOperationException

The operation is invalid.

InvalidPasswordException

The password is invalid.

LimitExceededException

The maximum of 100,000 folders under the parent folder has been exceeded.

ProhibitedStateException

The specified document version is not in the INITIALIZED state.

RemoveAllResourcePermissionsError

RemoveResourcePermissionError

RequestedEntityTooLargeException

The response is too large to return. The request must include a filter to reduce the size of the response.

ResourceAlreadyCheckedOutException

The resource is already checked out.

ServiceUnavailableException

One or more of the dependencies is unavailable.

StorageLimitExceededException

The storage limit has been exceeded.

StorageLimitWillExceedException

The storage limit will be exceeded.

TooManyLabelsException

The limit has been reached on the number of labels for the specified

TooManySubscriptionsException

You've reached the limit on the number of subscriptions for the WorkDocs

UnauthorizedOperationException

The operation is not permitted.

UnauthorizedResourceAccessException

The caller does not have access to perform the action on the resource.

UpdateDocumentError

UpdateDocumentVersionError

UpdateFolderError

UpdateUserError

Enums

AbortDocumentVersionUploadErrorKind

ActivateUserErrorKind

AddResourcePermissionsErrorKind

CreateCommentErrorKind

CreateCustomMetadataErrorKind

CreateFolderErrorKind

CreateLabelsErrorKind

CreateNotificationSubscriptionErrorKind

CreateUserErrorKind

DeactivateUserErrorKind

DeleteCommentErrorKind

DeleteCustomMetadataErrorKind

DeleteDocumentErrorKind

DeleteFolderContentsErrorKind

DeleteFolderErrorKind

DeleteLabelsErrorKind

DeleteNotificationSubscriptionErrorKind

DeleteUserErrorKind

DescribeActivitiesErrorKind

DescribeCommentsErrorKind

DescribeDocumentVersionsErrorKind

DescribeFolderContentsErrorKind

DescribeGroupsErrorKind

DescribeNotificationSubscriptionsErrorKind

DescribeResourcePermissionsErrorKind

DescribeRootFoldersErrorKind

DescribeUsersErrorKind

GetCurrentUserErrorKind

GetDocumentErrorKind

GetDocumentPathErrorKind

GetDocumentVersionErrorKind

GetFolderErrorKind

GetFolderPathErrorKind

GetResourcesErrorKind

InitiateDocumentVersionUploadErrorKind

RemoveAllResourcePermissionsErrorKind

RemoveResourcePermissionErrorKind

UpdateDocumentErrorKind

UpdateDocumentVersionErrorKind

UpdateFolderErrorKind

UpdateUserErrorKind