Module aws_sdk_workdocs::operation

Structs

AbortDocumentVersionUpload

Aborts the upload of the specified document version that was previously initiated

ActivateUser

Activates the specified user. Only active users can access Amazon

AddResourcePermissions

Creates a set of permissions for the specified folder or document. The resource

CreateComment

Adds a new comment to the specified document version.

CreateCustomMetadata

Adds one or more custom properties to the specified resource (a folder, document,

CreateFolder

Creates a folder with the specified name and parent folder.

CreateLabels

Adds the specified list of labels to the given resource (a document or

CreateNotificationSubscription

Configure Amazon WorkDocs to use Amazon SNS notifications. The endpoint receives a

CreateUser

Creates a user in a Simple AD or Microsoft AD directory. The status of a newly

DeactivateUser

Deactivates the specified user, which revokes the user's access to Amazon

DeleteComment

Deletes the specified comment from the document version.

DeleteCustomMetadata

Deletes custom metadata from the specified resource.

DeleteDocument

Permanently deletes the specified document and its associated metadata.

DeleteFolder

Permanently deletes the specified folder and its contents.

DeleteFolderContents

Deletes the contents of the specified folder.

DeleteLabels

Deletes the specified list of labels from a resource.

DeleteNotificationSubscription

Deletes the specified subscription from the specified organization.

DeleteUser

Deletes the specified user from a Simple AD or Microsoft AD directory.

DescribeActivities

Describes the user activities in a specified time period.

DescribeComments

List all the comments for the specified document version.

DescribeDocumentVersions

Retrieves the document versions for the specified document.

DescribeFolderContents

Describes the contents of the specified folder, including its documents and

DescribeGroups

Describes the groups specified by the query. Groups are defined by the underlying

DescribeNotificationSubscriptions

Lists the specified notification subscriptions.

DescribeResourcePermissions

Describes the permissions of a specified resource.

DescribeRootFolders

Describes the current user's special folders; the RootFolder and the

DescribeUsers

Describes the specified users. You can describe all users or filter the results

GetCurrentUser

Retrieves details of the current user for whom the authentication token was

GetDocument

Retrieves details of a document.

GetDocumentPath

Retrieves the path information (the hierarchy from the root folder) for the

GetDocumentVersion

Retrieves version metadata for the specified document.

GetFolder

Retrieves the metadata of the specified folder.

GetFolderPath

Retrieves the path information (the hierarchy from the root folder) for the

GetResources

Retrieves a collection of resources, including folders and documents. The only

InitiateDocumentVersionUpload

Creates a new document object and version object.

RemoveAllResourcePermissions

Removes all the permissions from the specified resource.

RemoveResourcePermission

Removes the permission for the specified principal from the specified

UpdateDocument

Updates the specified attributes of a document. The user must have access to both

UpdateDocumentVersion

Changes the status of the document version to ACTIVE.

UpdateFolder

Updates the specified attributes of the specified folder. The user must have access

UpdateUser

Updates the specified attributes of the specified user, and grants or revokes